Support custom documentation extensions
Some companies that are running Gerrit have some custom workflows and
tools around Gerrit. Those custom workflows and tools need to be
described and documented. Normally this happens outside of Gerrit on
some Wiki space.

For users it would be convinient to have this documentation directly
in Gerrit. The x-docs plugin allows to add a menu item under the
'Documentation' menu entry for such custom documentation. This
documentation is directly served from a Gerrit project (from which
project it is served is configured in the plugin). This has the
advantage that the development of the documentation is hosted in
Gerrit and you have version control and a code review process for it,
which also allows normal users to contribute to it.

The documentation has to be provided as Markdown files in the
configured Gerrit project. As entry point an index.md must be
available. If changes are merged into that project the served
documentation is immediatly providing the new content.

Change-Id: I442b6426c96761ffd4a76bd33beaddc44e6ceaf8
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>