Cache project documentation
The usage of a class that is declared in the plugin as key in a
persistent cache does not properly work. On loading from the cache
this fails with a ClassNotFoundException for the key while
deserializing. This is why we use String as key in the cache and not
the XDocResourceKey class.

Change-Id: Ie55fee77188b7005606cdb79a39c9c40e1a5329f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>