Register a project menu item that links to the project documentation
If a project is selected a 'Projects' > 'Readme' menu item appears
that links to the 'README.md' of the project (the one from the branch
to which HEAD points).

This menu item is also shown for projects that don't have a
'README.md'. In this case clicking on the menu items results in a
'Not Found' error.

Change-Id: I0b65a9db5fc8fd0e735ec1a19a07754960de4c16
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>