Allow to configure the documentation index file per project
At the moment a README.md file in the project root serves as entry
point to the project documentation and the documentation links in the
web UI link to this file. This change makes the index file
configurable per project, e.g. it could be set to
'Documentation/README.md' or 'index.md'. This feature will be
important in future when the plugin supports also other syntaxes, e.g.
a project may use AsciiDoc for their project documentation and then
wants to use 'README.adoc' as entry point for the project
documentation. The configuration is inherited from the parent
project. 'README.md' stays the default if nothing is configured.

Change-Id: Ie9df03b44adc85cc2ce38713a01075690b11b837
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>