Make it configurable which mime types should be handled as project doc
At the moment it is hard-coded that "*.md" files are handled as
Markdown project documentation. Instead one can now define in the
global plugin configuration which mime types should be handled by the
Markdown formatter:

[formatter "MARKDOWN"]
  mimeType = text/x-markdown

A mapping from file extensions to mime types already exists and
doesn't need to be reimplemented, instead just rely on the mime type.

Multiple mime types may be specified for a formatter. A formatter for
which no mime type is specified is disabled.

An init step creates the basic plugin configuration so that the
default mime type for a formatter is automatically set.

At the moment there is only a single formatter, the formatter for
Markdown, but follow-up changes will add formatters for other
syntaxes, such as Asciidoc.

Change-Id: I0f89463f89baaac962876a166e72ff8e26865280
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>