Add formatter for plain text project documentation
This adds a formatter that can display project documentation that is
written in plain "*.txt" files. While the preview functionality for
such files is of no use (you already see the plain text in the code
mirror editor) being able to browse plain text files has a value. E.g
it makes it possible to link to plain text files (or other resources
which could be viewed as plain text, e.g. script files) from within
project documentation that is written in Markdown. Also projects may
decide to have their project documentation completely as plain text
files. In this case the index file for the project could e.g. be
configured to "readme.txt" and a link to it would appear in the
project list.

This is also an example of how new formatters can be added.

Change-Id: I06b962c95f5d763b6009fe42534cf8ae73f369e5
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>