Add an asciidoctor formatter for rendering asciidoc
Asciidoc doesn't have an own mime type and a standard file extension.
Often '*.adoc' with mime type 'text/x-asciidoc' is used, but we don't
have this association in Gerrit. This means for now you can make use
of the asciidoctor formatter only if you configure it for another mime
type, e.g. have the asciidoc documentation in '*.txt' files and
configure the asciidoctor formatter to handle the 'text/plain' mime
type:

  [formatter "ASCIIDOCTOR"]
    mimeType = text/plain

Change-Id: I6a5fe4e6d4e97c3b0ea4929a015116cb3820f814
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>