[RFC] Support configuring formatters for file extensions
At the moment formatters can only be configured to handle files with a
certain mime type. The problem is that some formats don't have an own
standardized mime type. E.g. Asciidoc files often use the file
extension "*.adoc" but the mime type "text/x-asciidoc" for it is not
fully established and Gerrit doesn't recognize it.

Change-Id: I461d3c3e5b0d7f250705161b39a6b2d4e107275b
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>