Support configuring formatters for file extensions
At the moment formatters can only be configured to handle files with a
certain mime type. The problem is that some formats don't have an own
standardized mime type. E.g. Asciidoc files often use the file
extension "*.adoc" but the mime type "text/x-asciidoc" for it is not
fully established and Gerrit doesn't recognize it.

Change-Id: Ifedbe393a66f3e53ad3b3c67efcb98a1208e303c
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>