Allow admins to disable certain formatters
If a formatter is found to be a security risk it must be possible to
disable it globally. Add a new configuration parameter to disable a
formatter:

  [formatter "ASCIIDOC"]
    enabled = false

Before I5b095ab4c it was possible to disable formatters by not
specifying any mime type or file extension for a formatter, but since
I5b095ab4c projects can now overwrite this. This is why we need an
explicit way to disable a formatter now.

Change-Id: I1ea11e58cf7a75d1befe3149acae0d44b0e550a6
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>