Add Gerrit screen for showing project documentation
Add a new Gerrit screen that shows the rendered project documentation
and some information about the viewed file in a header above it.
For now the header only shows the project, the file name and the
revision. In the future more functionality can be added here, e.g.
switch to edit mode or browsing of the file hierarchy.

The screen is available under
  #/x/project/<project-name>/<file-name>
and
  #/x/project/<project-name>/rev/<revision>/<file-name>

Change-Id: Ie9025cab92bdeea4d925c11fd8f60f5dc512e955
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>