Allow projects to define own CSS for Markdown files
Projects may have different preferences of how their Markdown
documentation should be formatted. By default the standard Gerrit CSS
for Markdown files is used, but now projects can override it by custom
CSS that is stored in the project's refs/meta/config branch as
'xdocs/markdown.css'.

The current revision of the refs/meta/config branch is included in the
cache key of the project documentation so that the project
documentation is newly formatted if there was a change in this branch.
This ensures that the new CSS is immediately active.

Custom CSS is not inherited from parent projects.

Change-Id: I909ed431a103aa6a6561ae8fc1d558bfb57f6af8
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>