Configure ASCIIDOCTOR formatter
Set options and attributes for asciidoc generation, include CSS and
set the safe mode to SECURE.

A new configuration parameter allows to control whether a Table Of
Contents should be included at the beginning of each document. For now
this is a global configuration, but a follow-up change should make
this configurable per project.

Change-Id: I51faf84e8edd6a6e515e8f2693acd5d45929501f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>