Show proper error message in UI when document cannot be loaded
Instead of showing a broken iframe, display a proper error message
when a document cannot be loaded.

This is a preparation step to add special handling if the document is
not found, e.g. in this case we may want to display a 'Create' button
to create the missing document.

Change-Id: Ic0b7383da84f2351c24ae49eccc7c8f6fc421c9d
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>