Allow to specify priorities for formatters
If two formatters can handle a file it is currently undefined which
formatter is taken.

Example:
For file 'Documentation/index.txt' and the following configuration it
undefined if the "ASCIIDOCTOR" or "PLAIN_TEXT" formatter is used:

  [formatter "ASCIIDOCTOR"]
    ext = txt
    prefix = Documentation/
  [formatter "PLAIN_TEXT"]
    mimeType = text/plain

This can now be defined by specifying a formatter priority:

  [formatter "ASCIIDOCTOR"]
    ext = txt
    prefix = Documentation/
    prio = 2
  [formatter "PLAIN_TEXT"]
    mimeType = text/plain
    prio = 1

The formatter with the higher priority wins, 0 is the default priority.

Change-Id: I29ba5bf58d00b3591029f203abd7f59893a2ab39
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>