Return null as web link if no web link should be displayed
So far a web link without URL was returned to indicate that no link
should be displayed, but with I28b98c25a this now results in a warning
in the log that we don't want to see.

Also the web links targets are now exposed as an inner class of the
WebLink interface. Adapt the plugin code to this change.

Change-Id: Ifd10df61ad80bf792348b5bd99aa2e5cf5288635
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>