Support CSS themes
This allows administrators to configure multiple default CSS themes
and projects can simply choose which one should be used. Also projects
can define CSS themes, e.g. if a new CSS theme is developed, this can
be done while still using the old theme and then switch to the new
theme when it is ready.

Change-Id: Ia7a36b8367d9d09da3775e9cb8e4e04122a319fe
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>