Allow to configure a default ref for loading the documentation
So far if a revision was not specified, the documentation was always
loaded from the branch to which HEAD points. Now a new configuration
parameter allows to configure the default ref from which the
documentation is loaded. E.g. it could be set to 'master' or to a tag
that marks the latest release 'refs/tags/v2.9.1'

Change-Id: I6a22e6de5d2593e87971012d6514b28f1af35aad
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>