Update patch set 2
Patch Set 2: Code-Review+1

All good, works for me.
However the caching caught me in doubt ... seems to be a bit aggressive and even when I changed the settings in All-Projects/refs/meta/config the old page was still served.

However this is not going to change so often, so I presume it is fine :-)

Patch-set: 2
Label: Code-Review=+1