Reload documents if meta data of any parent project is changed
Meta data changes in a parent project may affect the rendering of
documents. Hence documents should be reloaded if there was a change in
the refs/meta/config branch of any parent project. To achieve this a
hash computed out of the revision IDs of the refs/meta/config branches
of all parent project is included into the document cache key. As
result there is no cache hit when the meta data of a parent project
was changed and the document is reloaded.

This will also enable the implementation of inheritable CSS.

Change-Id: I9dabffabc8a003ef42a21e280eeba355462c2a8a
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>