Add support for stream formatters
The current formatters expect the file content as String, which is
good for small files where the whole content will be reformatted as
HTML. It is bad for formatters that generate a small HTML page for a
large file, e.g. a ZipFormatter could show a list with the files that
are contained in a zip file. In this case it is unneeded and
resource-wasting to load the whole content into memory. To support
such formatters a new StreamFormatter interface is introduced. For
StreamFormatters there is no macro replacement done.

Change-Id: I182b6a735b4e7aeb1ab7705c760144278294f81f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>