Add formatter for .docx, .pptx and .xlsx files
This allows to look at MS Office documents. Please note that this
formatter is not working for .doc, .ppt and .xls files.

Docx4J tries to load some resources dynamically. This fails if the
Gerrit core classloader is used since it doesn't see the resources
that are contained in the plugin jar. To make the resource loading
work the DocxFormatter must set a context classloader on the current
thread.

Change-Id: Iffe6bf4f469fc237cbb79d37926e80d61e337649
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>