Refactor XDocServlet: Move image checks into helper methods
The code in XDocServlet grew over time. Due to this it became less
readable and by now the control flow is difficult to understand.

This is the first change in a series that refactors the XDocServlet to
make the code better readable again. The refactoring is split into
small steps so that reviewers can easier follow what is done.

Change-Id: I45f556ca57622038d08c1b41efd182eddaf9c9b9
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>