Add navigation links to side-by-side and unified diff screens
The navigation links allow to navigate to the previous/next file in
the change and back to the change screen. If a formatter is available
for the previous/next file, the diff is again shown in the HTML diff
view, otherwise it goes back to the normal side-by-side/unified diff
view.

Change-Id: I1ba63b90110e8d41e0ab10083b15ae92bda58134
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>