Add diff web links to switch from normal diff to preview diff
This makes the preview diff screens accessible for user. If a preview
is available for a file then the side-by-side/unified diff screen
shows an icon to switch to the side-by-side/unified preview diff
screen. Before this change the URL for the preview diff screens had
to be typed manually.

Icons to navigate back from the preview diff screens to the normal
diff screens will be added in a follow-up change.

Change-Id: Ic88beef1f8c2d1e5c7f260f38cfe56702a92e759
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>