Move CSS handling into FormatterUtil to make it reuseable
Which CSS is applied to a document depends on a number of
configuration paramters and on the availability of default CSS,
inherited CSS and project-specific CSS. Basically each CSS-aware
formatter would need to do the same computation to apply the correct
CSS. Having the code for it in a common place avoids reimplementation
of this logic again and again.

Please note that the CSS handling in the MarkdownFormatter is special
since it wants to reuse the default CSS that is packaged in Gerrit
core. This is why the MarkdownFormatter doesn't use the CSS function
from FormatterUtil.

Change-Id: I6fce8b6912ef8242c26a003438f1fe9038ba4150
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>