Update patch set 1
Patch Set 1:

(1 comment)

> > * <missing web link> click preview icon and jump directly to
 > > x-docs-sbs
 > >
 > > What we need is to extend inline edit to support webl links.
 > 
 > +1

What I would really love to achieve is to have a split screen, where we have an editor an the left pane and the preview on the right pane and updating the preview on each save. But not sure yet how to get this.

Patch-set: 1
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0