Add formatter to render images
This allows to compare old and new image in the preview diff screens.

The image formatter is only used if a diff is requested. If no diff is
requested the plain images are returned. This is important to render
images that are embedded in project documentation, e.g. in Markdown or
Asciidoc files.

Change-Id: I17edc14a42c74100d0588ccbbf601a20807e356a
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>