Do not encode file path in URL
If the file path is encoded relative links cannot be resolved, e.g.
a file 'foo/bar.md' which contains a relative link to 'foo/other.md'
as '[other](other.md)' would be resolved to 'other.md' instead of
'foo/other.md' (because '/' in the URL of 'foo/bar.md' is encoded as
'%2F').

Change-Id: I6eecdec621caa88893f98f5b1330138f861b1f8c
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>