Fix rendering of embedded images in preview diff
For embedded images the XDocServlet should return the plain image
file, but for the preview of images it should use the ImageFormatter
and return the formatted image (e.g. with green background if added,
with red background if deleted). Add a URL parameter 'formatImage'
that decides if the ImageFormatter is used. If not given the plain
image file is returned.

Change-Id: I3a6ee6fb2738603aa45172d2ae7d5339945b7802
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>