Update patch set 7
Patch Set 7:

> Thanks.

So, unzipping non-stamped jar version and zipping it again is failin, because, there are some directories inside some deps, that cannot be read. That sucks:

  ./org/eclipse/core/runtime:
  total 52
  drwxr-xr-x 2 davido users 4096 Jan  1  1980 .
  drwxr-xr-x 3 davido users 4096 Jan  1  1980 ..
  ---------- 1 davido users 1352 Jan  1  1980 Assert.class
  ---------- 1 davido users  484 Jan  1  1980 AssertionFailedException.class

It's starting to fail, when we are adding this dep:

  "@daisydiff//jar",

and its deps. So something is broken there.

Patch-set: 7
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>