Fix build with buck in Gerrit tree
Since previous version, buck build in Gerrit tree fails
because no rule is provided to generate asciidoctor licence.
The error is:
No rule found when resolving target //lib:LICENSE-asciidoctor

As the licence is available in the lib folder, this fix
simply toggles on the "local_license" option for asciidoctor.

It seems buck build is deprecated now but this fix may
interest people still using Gerrit 2.11

Change-Id: I939eccaa6434602c9ea8b5e479136b7b93234fc9